services/kms/requirements.txt
fastapi
uvicorn[standard]
pycryptodome